Decommission config and move to yaml, add default icon support, fix issue with saving
