--- modules/backend/config/icons.php
@@ -1,1583 +1,0 @@
-<?php
-
-/**
- * This is a temporary solution
- * Ideally we should update FontAwesome, and generate this somewhere else from all the available icons.
- * This way there will be no manual intervention to keep the icons updated for the Iconpicker.
- */
-
-return [
-    [
-        "id" => "far",
-        "title" => "Font Awesome Regular",
-        "prefix" => "far icon-",
-        "listicon" => "far icon-circle",
-        "icons" => [
-            "far icon-address-book",
-            "far icon-address-card",
-            "far icon-angry",
-            "far icon-arrow-alt-circle-down",
-            "far icon-arrow-alt-circle-left",
-            "far icon-arrow-alt-circle-right",
-            "far icon-arrow-alt-circle-up",
-            "far icon-bell",
-            "far icon-bell-slash",
-            "far icon-bookmark",
-            "far icon-building",
-            "far icon-calendar",
-            "far icon-calendar-alt",
-            "far icon-calendar-check",
-            "far icon-calendar-minus",
-            "far icon-calendar-plus",
-            "far icon-calendar-times",
-            "far icon-caret-square-down",
-            "far icon-caret-square-left",
-            "far icon-caret-square-right",
-            "far icon-caret-square-up",
-            "far icon-chart-bar",
-            "far icon-check-circle",
-            "far icon-check-square",
-            "far icon-circle",
-            "far icon-clipboard",
-            "far icon-clock",
-            "far icon-clone",
-            "far icon-closed-captioning",
-            "far icon-comment",
-            "far icon-comment-alt",
-            "far icon-comment-dots",
-            "far icon-comments",
-            "far icon-compass",
-            "far icon-copy",
-            "far icon-copyright",
-            "far icon-credit-card",
-            "far icon-dizzy",
-            "far icon-dot-circle",
-            "far icon-edit",
-            "far icon-envelope",
-            "far icon-envelope-open",
-            "far icon-eye",
-            "far icon-eye-slash",
-            "far icon-file",
-            "far icon-file-alt",
-            "far icon-file-archive",
-            "far icon-file-audio",
-            "far icon-file-code",
-            "far icon-file-excel",
-            "far icon-file-image",
-            "far icon-file-pdf",
-            "far icon-file-powerpoint",
-            "far icon-file-video",
-            "far icon-file-word",
-            "far icon-flag",
-            "far icon-flushed",
-            "far icon-folder",
-            "far icon-folder-open",
-            "far icon-frown",
-            "far icon-frown-open",
-            "far icon-futbol",
-            "far icon-gem",
-            "far icon-grimace",
-            "far icon-grin",
-            "far icon-grin-alt",
-            "far icon-grin-beam",
-            "far icon-grin-beam-sweat",
-            "far icon-grin-hearts",
-            "far icon-grin-squint",
-            "far icon-grin-squint-tears",
-            "far icon-grin-stars",
-            "far icon-grin-tears",
-            "far icon-grin-tongue",
-            "far icon-grin-tongue-squint",
-            "far icon-grin-tongue-wink",
-            "far icon-grin-wink",
-            "far icon-hand-lizard",
-            "far icon-hand-paper",
-            "far icon-hand-peace",
-            "far icon-hand-point-down",
-            "far icon-hand-point-left",
-            "far icon-hand-point-right",
-            "far icon-hand-point-up",
-            "far icon-hand-pointer",
-            "far icon-hand-rock",
-            "far icon-hand-scissors",
-            "far icon-hand-spock",
-            "far icon-handshake",
-            "far icon-hdd",
-            "far icon-heart",
-            "far icon-hospital",
-            "far icon-hourglass",
-            "far icon-id-badge",
-            "far icon-id-card",
-            "far icon-image",
-            "far icon-images",
-            "far icon-keyboard",
-            "far icon-kiss",
-            "far icon-kiss-beam",
-            "far icon-kiss-wink-heart",
-            "far icon-laugh",
-            "far icon-laugh-beam",
-            "far icon-laugh-squint",
-            "far icon-laugh-wink",
-            "far icon-lemon",
-            "far icon-life-ring",
-            "far icon-lightbulb",
-            "far icon-list-alt",
-            "far icon-map",
-            "far icon-meh",
-            "far icon-meh-blank",
-            "far icon-meh-rolling-eyes",
-            "far icon-minus-square",
-            "far icon-money-bill-alt",
-            "far icon-moon",
-            "far icon-newspaper",
-            "far icon-object-group",
-            "far icon-object-ungroup",
-            "far icon-paper-plane",
-            "far icon-pause-circle",
-            "far icon-play-circle",
-            "far icon-plus-square",
-            "far icon-question-circle",
-            "far icon-registered",
-            "far icon-sad-cry",
-            "far icon-sad-tear",
-            "far icon-save",
-            "far icon-share-square",
-            "far icon-smile",
-            "far icon-smile-beam",
-            "far icon-smile-wink",
-            "far icon-snowflake",
-            "far icon-square",
-            "far icon-star",
-            "far icon-star-half",
-            "far icon-sticky-note",
-            "far icon-stop-circle",
-            "far icon-sun",
-            "far icon-surprise",
-            "far icon-thumbs-down",
-            "far icon-thumbs-up",
-            "far icon-times-circle",
-            "far icon-tired",
-            "far icon-trash-alt",
-            "far icon-user",
-            "far icon-user-circle",
-            "far icon-window-close",
-            "far icon-window-maximize",
-            "far icon-window-minimize",
-            "far icon-window-restore"
-        ]
-    ],
-    [
-        "id" => "fas",
-        "title" => "Font Awesome Solid",
-        "prefix" => "fas icon-",
-        "listicon" => "fas icon-circle",
-        "icons" => [
-            "fas icon-ad",
-            "fas icon-address-book",
-            "fas icon-address-card",
-            "fas icon-adjust",
-            "fas icon-air-freshener",
-            "fas icon-align-center",
-            "fas icon-align-justify",
-            "fas icon-align-left",
-            "fas icon-align-right",
-            "fas icon-allergies",
-            "fas icon-ambulance",
-            "fas icon-american-sign-language-interpreting",
-            "fas icon-anchor",
-            "fas icon-angle-double-down",
-            "fas icon-angle-double-left",
-            "fas icon-angle-double-right",
-            "fas icon-angle-double-up",
-            "fas icon-angle-down",
-            "fas icon-angle-left",
-            "fas icon-angle-right",
-            "fas icon-angle-up",
-            "fas icon-angry",
-            "fas icon-ankh",
-            "fas icon-apple-alt",
-            "fas icon-archive",
-            "fas icon-archway",
-            "fas icon-arrow-alt-circle-down",
-            "fas icon-arrow-alt-circle-left",
-            "fas icon-arrow-alt-circle-right",
-            "fas icon-arrow-alt-circle-up",
-            "fas icon-arrow-circle-down",
-            "fas icon-arrow-circle-left",
-            "fas icon-arrow-circle-right",
-            "fas icon-arrow-circle-up",
-            "fas icon-arrow-down",
-            "fas icon-arrow-left",
-            "fas icon-arrow-right",
-            "fas icon-arrow-up",
-            "fas icon-arrows-alt",
-            "fas icon-arrows-alt-h",
-            "fas icon-arrows-alt-v",
-            "fas icon-assistive-listening-systems",
-            "fas icon-asterisk",
-            "fas icon-at",
-            "fas icon-atlas",
-            "fas icon-atom",
-            "fas icon-audio-description",
-            "fas icon-award",
-            "fas icon-baby",
-            "fas icon-baby-carriage",
-            "fas icon-backspace",
-            "fas icon-backward",
-            "fas icon-bacon",
-            "fas icon-bahai",
-            "fas icon-balance-scale",
-            "fas icon-balance-scale-left",
-            "fas icon-balance-scale-right",
-            "fas icon-ban",
-            "fas icon-band-aid",
-            "fas icon-barcode",
-            "fas icon-bars",
-            "fas icon-baseball-ball",
-            "fas icon-basketball-ball",
-            "fas icon-bath",
-            "fas icon-battery-empty",
-            "fas icon-battery-full",
-            "fas icon-battery-half",
-            "fas icon-battery-quarter",
-            "fas icon-battery-three-quarters",
-            "fas icon-bed",
-            "fas icon-beer",
-            "fas icon-bell",
-            "fas icon-bell-slash",
-            "fas icon-bezier-curve",
-            "fas icon-bible",
-            "fas icon-bicycle",
-            "fas icon-biking",
-            "fas icon-binoculars",
-            "fas icon-biohazard",
-            "fas icon-birthday-cake",
-            "fas icon-blender",
-            "fas icon-blender-phone",
-            "fas icon-blind",
-            "fas icon-blog",
-            "fas icon-bold",
-            "fas icon-bolt",
-            "fas icon-bomb",
-            "fas icon-bone",
-            "fas icon-bong",
-            "fas icon-book",
-            "fas icon-book-dead",
-            "fas icon-book-medical",
-            "fas icon-book-open",
-            "fas icon-book-reader",
-            "fas icon-bookmark",
-            "fas icon-border-all",
-            "fas icon-border-none",
-            "fas icon-border-style",
-            "fas icon-bowling-ball",
-            "fas icon-box",
-            "fas icon-box-open",
-            "fas icon-boxes",
-            "fas icon-braille",
-            "fas icon-brain",
-            "fas icon-bread-slice",
-            "fas icon-briefcase",
-            "fas icon-briefcase-medical",
-            "fas icon-broadcast-tower",
-            "fas icon-broom",
-            "fas icon-brush",
-            "fas icon-bug",
-            "fas icon-building",
-            "fas icon-bullhorn",
-            "fas icon-bullseye",
-            "fas icon-burn",
-            "fas icon-bus",
-            "fas icon-bus-alt",
-            "fas icon-business-time",
-            "fas icon-calculator",
-            "fas icon-calendar",
-            "fas icon-calendar-alt",
-            "fas icon-calendar-check",
-            "fas icon-calendar-day",
-            "fas icon-calendar-minus",
-            "fas icon-calendar-plus",
-            "fas icon-calendar-times",
-            "fas icon-calendar-week",
-            "fas icon-camera",
-            "fas icon-camera-retro",
-            "fas icon-campground",
-            "fas icon-candy-cane",
-            "fas icon-cannabis",
-            "fas icon-capsules",
-            "fas icon-car",
-            "fas icon-car-alt",
-            "fas icon-car-battery",
-            "fas icon-car-crash",
-            "fas icon-car-side",
-            "fas icon-caravan",
-            "fas icon-caret-down",
-            "fas icon-caret-left",
-            "fas icon-caret-right",
-            "fas icon-caret-square-down",
-            "fas icon-caret-square-left",
-            "fas icon-caret-square-right",
-            "fas icon-caret-square-up",
-            "fas icon-caret-up",
-            "fas icon-carrot",
-            "fas icon-cart-arrow-down",
-            "fas icon-cart-plus",
-            "fas icon-cash-register",
-            "fas icon-cat",
-            "fas icon-certificate",
-            "fas icon-chair",
-            "fas icon-chalkboard",
-            "fas icon-chalkboard-teacher",
-            "fas icon-charging-station",
-            "fas icon-chart-area",
-            "fas icon-chart-bar",
-            "fas icon-chart-line",
-            "fas icon-chart-pie",
-            "fas icon-check",
-            "fas icon-check-circle",
-            "fas icon-check-double",
-            "fas icon-check-square",
-            "fas icon-cheese",
-            "fas icon-chess",
-            "fas icon-chess-bishop",
-            "fas icon-chess-board",
-            "fas icon-chess-king",
-            "fas icon-chess-knight",
-            "fas icon-chess-pawn",
-            "fas icon-chess-queen",
-            "fas icon-chess-rook",
-            "fas icon-chevron-circle-down",
-            "fas icon-chevron-circle-left",
-            "fas icon-chevron-circle-right",
-            "fas icon-chevron-circle-up",
-            "fas icon-chevron-down",
-            "fas icon-chevron-left",
-            "fas icon-chevron-right",
-            "fas icon-chevron-up",
-            "fas icon-child",
-            "fas icon-church",
-            "fas icon-circle",
-            "fas icon-circle-notch",
-            "fas icon-city",
-            "fas icon-clinic-medical",
-            "fas icon-clipboard",
-            "fas icon-clipboard-check",
-            "fas icon-clipboard-list",
-            "fas icon-clock",
-            "fas icon-clone",
-            "fas icon-closed-captioning",
-            "fas icon-cloud",
-            "fas icon-cloud-download-alt",
-            "fas icon-cloud-meatball",
-            "fas icon-cloud-moon",
-            "fas icon-cloud-moon-rain",
-            "fas icon-cloud-rain",
-            "fas icon-cloud-showers-heavy",
-            "fas icon-cloud-sun",
-            "fas icon-cloud-sun-rain",
-            "fas icon-cloud-upload-alt",
-            "fas icon-cocktail",
-            "fas icon-code",
-            "fas icon-code-branch",
-            "fas icon-coffee",
-            "fas icon-cog",
-            "fas icon-cogs",
-            "fas icon-coins",
-            "fas icon-columns",
-            "fas icon-comment",
-            "fas icon-comment-alt",
-            "fas icon-comment-dollar",
-            "fas icon-comment-dots",
-            "fas icon-comment-medical",
-            "fas icon-comment-slash",
-            "fas icon-comments",
-            "fas icon-comments-dollar",
-            "fas icon-compact-disc",
-            "fas icon-compass",
-            "fas icon-compress",
-            "fas icon-compress-alt",
-            "fas icon-compress-arrows-alt",
-            "fas icon-concierge-bell",
-            "fas icon-cookie",
-            "fas icon-cookie-bite",
-            "fas icon-copy",
-            "fas icon-copyright",
-            "fas icon-couch",
-            "fas icon-credit-card",
-            "fas icon-crop",
-            "fas icon-crop-alt",
-            "fas icon-cross",
-            "fas icon-crosshairs",
-            "fas icon-crow",
-            "fas icon-crown",
-            "fas icon-crutch",
-            "fas icon-cube",
-            "fas icon-cubes",
-            "fas icon-cut",
-            "fas icon-database",
-            "fas icon-deaf",
-            "fas icon-democrat",
-            "fas icon-desktop",
-            "fas icon-dharmachakra",
-            "fas icon-diagnoses",
-            "fas icon-dice",
-            "fas icon-dice-d20",
-            "fas icon-dice-d6",
-            "fas icon-dice-five",
-            "fas icon-dice-four",
-            "fas icon-dice-one",
-            "fas icon-dice-six",
-            "fas icon-dice-three",
-            "fas icon-dice-two",
-            "fas icon-digital-tachograph",
-            "fas icon-directions",
-            "fas icon-divide",
-            "fas icon-dizzy",
-            "fas icon-dna",
-            "fas icon-dog",
-            "fas icon-dollar-sign",
-            "fas icon-dolly",
-            "fas icon-dolly-flatbed",
-            "fas icon-donate",
-            "fas icon-door-closed",
-            "fas icon-door-open",
-            "fas icon-dot-circle",
-            "fas icon-dove",
-            "fas icon-download",
-            "fas icon-drafting-compass",
-            "fas icon-dragon",
-            "fas icon-draw-polygon",
-            "fas icon-drum",
-            "fas icon-drum-steelpan",
-            "fas icon-drumstick-bite",
-            "fas icon-dumbbell",
-            "fas icon-dumpster",
-            "fas icon-dumpster-fire",
-            "fas icon-dungeon",
-            "fas icon-edit",
-            "fas icon-egg",
-            "fas icon-eject",
-            "fas icon-ellipsis-h",
-            "fas icon-ellipsis-v",
-            "fas icon-envelope",
-            "fas icon-envelope-open",
-            "fas icon-envelope-open-text",
-            "fas icon-envelope-square",
-            "fas icon-equals",
-            "fas icon-eraser",
-            "fas icon-ethernet",
-            "fas icon-euro-sign",
-            "fas icon-exchange-alt",
-            "fas icon-exclamation",
-            "fas icon-exclamation-circle",
-            "fas icon-exclamation-triangle",
-            "fas icon-expand",
-            "fas icon-expand-alt",
-            "fas icon-expand-arrows-alt",
-            "fas icon-external-link-alt",
-            "fas icon-external-link-square-alt",
-            "fas icon-eye",
-            "fas icon-eye-dropper",
-            "fas icon-eye-slash",
-            "fas icon-fan",
-            "fas icon-fast-backward",
-            "fas icon-fast-forward",
-            "fas icon-fax",
-            "fas icon-feather",
-            "fas icon-feather-alt",
-            "fas icon-female",
-            "fas icon-fighter-jet",
-            "fas icon-file",
-            "fas icon-file-alt",
-            "fas icon-file-archive",
-            "fas icon-file-audio",
-            "fas icon-file-code",
-            "fas icon-file-contract",
-            "fas icon-file-csv",
-            "fas icon-file-download",
-            "fas icon-file-excel",
-            "fas icon-file-export",
-            "fas icon-file-image",
-            "fas icon-file-import",
-            "fas icon-file-invoice",
-            "fas icon-file-invoice-dollar",
-            "fas icon-file-medical",
-            "fas icon-file-medical-alt",
-            "fas icon-file-pdf",
-            "fas icon-file-powerpoint",
-            "fas icon-file-prescription",
-            "fas icon-file-signature",
-            "fas icon-file-upload",
-            "fas icon-file-video",
-            "fas icon-file-word",
-            "fas icon-fill",
-            "fas icon-fill-drip",
-            "fas icon-film",
-            "fas icon-filter",
-            "fas icon-fingerprint",
-            "fas icon-fire",
-            "fas icon-fire-alt",
-            "fas icon-fire-extinguisher",
-            "fas icon-first-aid",
-            "fas icon-fish",
-            "fas icon-fist-raised",
-            "fas icon-flag",
-            "fas icon-flag-checkered",
-            "fas icon-flag-usa",
-            "fas icon-flask",
-            "fas icon-flushed",
-            "fas icon-folder",
-            "fas icon-folder-minus",
-            "fas icon-folder-open",
-            "fas icon-folder-plus",
-            "fas icon-font",
-            "fas icon-football-ball",
-            "fas icon-forward",
-            "fas icon-frog",
-            "fas icon-frown",
-            "fas icon-frown-open",
-            "fas icon-funnel-dollar",
-            "fas icon-futbol",
-            "fas icon-gamepad",
-            "fas icon-gas-pump",
-            "fas icon-gavel",
-            "fas icon-gem",
-            "fas icon-genderless",
-            "fas icon-ghost",
-            "fas icon-gift",
-            "fas icon-gifts",
-            "fas icon-glass-cheers",
-            "fas icon-glass-martini",
-            "fas icon-glass-martini-alt",
-            "fas icon-glass-whiskey",
-            "fas icon-glasses",
-            "fas icon-globe",
-            "fas icon-globe-africa",
-            "fas icon-globe-americas",
-            "fas icon-globe-asia",
-            "fas icon-globe-europe",
-            "fas icon-golf-ball",
-            "fas icon-gopuram",
-            "fas icon-graduation-cap",
-            "fas icon-greater-than",
-            "fas icon-greater-than-equal",
-            "fas icon-grimace",
-            "fas icon-grin",
-            "fas icon-grin-alt",
-            "fas icon-grin-beam",
-            "fas icon-grin-beam-sweat",
-            "fas icon-grin-hearts",
-            "fas icon-grin-squint",
-            "fas icon-grin-squint-tears",
-            "fas icon-grin-stars",
-            "fas icon-grin-tears",
-            "fas icon-grin-tongue",
-            "fas icon-grin-tongue-squint",
-            "fas icon-grin-tongue-wink",
-            "fas icon-grin-wink",
-            "fas icon-grip-horizontal",
-            "fas icon-grip-lines",
-            "fas icon-grip-lines-vertical",
-            "fas icon-grip-vertical",
-            "fas icon-guitar",
-            "fas icon-h-square",
-            "fas icon-hamburger",
-            "fas icon-hammer",
-            "fas icon-hamsa",
-            "fas icon-hand-holding",
-            "fas icon-hand-holding-heart",
-            "fas icon-hand-holding-usd",
-            "fas icon-hand-lizard",
-            "fas icon-hand-middle-finger",
-            "fas icon-hand-paper",
-            "fas icon-hand-peace",
-            "fas icon-hand-point-down",
-            "fas icon-hand-point-left",
-            "fas icon-hand-point-right",
-            "fas icon-hand-point-up",
-            "fas icon-hand-pointer",
-            "fas icon-hand-rock",
-            "fas icon-hand-scissors",
-            "fas icon-hand-spock",
-            "fas icon-hands",
-            "fas icon-hands-helping",
-            "fas icon-handshake",
-            "fas icon-hanukiah",
-            "fas icon-hard-hat",
-            "fas icon-hashtag",
-            "fas icon-hat-cowboy",
-            "fas icon-hat-cowboy-side",
-            "fas icon-hat-wizard",
-            "fas icon-hdd",
-            "fas icon-heading",
-            "fas icon-headphones",
-            "fas icon-headphones-alt",
-            "fas icon-headset",
-            "fas icon-heart",
-            "fas icon-heart-broken",
-            "fas icon-heartbeat",
-            "fas icon-helicopter",
-            "fas icon-highlighter",
-            "fas icon-hiking",
-            "fas icon-hippo",
-            "fas icon-history",
-            "fas icon-hockey-puck",
-            "fas icon-holly-berry",
-            "fas icon-home",
-            "fas icon-horse",
-            "fas icon-horse-head",
-            "fas icon-hospital",
-            "fas icon-hospital-alt",
-            "fas icon-hospital-symbol",
-            "fas icon-hot-tub",
-            "fas icon-hotdog",
-            "fas icon-hotel",
-            "fas icon-hourglass",
-            "fas icon-hourglass-end",
-            "fas icon-hourglass-half",
-            "fas icon-hourglass-start",
-            "fas icon-house-damage",
-            "fas icon-hryvnia",
-            "fas icon-i-cursor",
-            "fas icon-ice-cream",
-            "fas icon-icicles",
-            "fas icon-icons",
-            "fas icon-id-badge",
-            "fas icon-id-card",
-            "fas icon-id-card-alt",
-            "fas icon-igloo",
-            "fas icon-image",
-            "fas icon-images",
-            "fas icon-inbox",
-            "fas icon-indent",
-            "fas icon-industry",
-            "fas icon-infinity",
-            "fas icon-info",
-            "fas icon-info-circle",
-            "fas icon-italic",
-            "fas icon-jedi",
-            "fas icon-joint",
-            "fas icon-journal-whills",
-            "fas icon-kaaba",
-            "fas icon-key",
-            "fas icon-keyboard",
-            "fas icon-khanda",
-            "fas icon-kiss",
-            "fas icon-kiss-beam",
-            "fas icon-kiss-wink-heart",
-            "fas icon-kiwi-bird",
-            "fas icon-landmark",
-            "fas icon-language",
-            "fas icon-laptop",
-            "fas icon-laptop-code",
-            "fas icon-laptop-medical",
-            "fas icon-laugh",
-            "fas icon-laugh-beam",
-            "fas icon-laugh-squint",
-            "fas icon-laugh-wink",
-            "fas icon-layer-group",
-            "fas icon-leaf",
-            "fas icon-lemon",
-            "fas icon-less-than",
-            "fas icon-less-than-equal",
-            "fas icon-level-down-alt",
-            "fas icon-level-up-alt",
-            "fas icon-life-ring",
-            "fas icon-lightbulb",
-            "fas icon-link",
-            "fas icon-lira-sign",
-            "fas icon-list",
-            "fas icon-list-alt",
-            "fas icon-list-ol",
-            "fas icon-list-ul",
-            "fas icon-location-arrow",
-            "fas icon-lock",
-            "fas icon-lock-open",
-            "fas icon-long-arrow-alt-down",
-            "fas icon-long-arrow-alt-left",
-            "fas icon-long-arrow-alt-right",
-            "fas icon-long-arrow-alt-up",
-            "fas icon-low-vision",
-            "fas icon-luggage-cart",
-            "fas icon-magic",
-            "fas icon-magnet",
-            "fas icon-mail-bulk",
-            "fas icon-male",
-            "fas icon-map",
-            "fas icon-map-marked",
-            "fas icon-map-marked-alt",
-            "fas icon-map-marker",
-            "fas icon-map-marker-alt",
-            "fas icon-map-pin",
-            "fas icon-map-signs",
-            "fas icon-marker",
-            "fas icon-mars",
-            "fas icon-mars-double",
-            "fas icon-mars-stroke",
-            "fas icon-mars-stroke-h",
-            "fas icon-mars-stroke-v",
-            "fas icon-mask",
-            "fas icon-medal",
-            "fas icon-medkit",
-            "fas icon-meh",
-            "fas icon-meh-blank",
-            "fas icon-meh-rolling-eyes",
-            "fas icon-memory",
-            "fas icon-menorah",
-            "fas icon-mercury",
-            "fas icon-meteor",
-            "fas icon-microchip",
-            "fas icon-microphone",
-            "fas icon-microphone-alt",
-            "fas icon-microphone-alt-slash",
-            "fas icon-microphone-slash",
-            "fas icon-microscope",
-            "fas icon-minus",
-            "fas icon-minus-circle",
-            "fas icon-minus-square",
-            "fas icon-mitten",
-            "fas icon-mobile",
-            "fas icon-mobile-alt",
-            "fas icon-money-bill",
-            "fas icon-money-bill-alt",
-            "fas icon-money-bill-wave",
-            "fas icon-money-bill-wave-alt",
-            "fas icon-money-check",
-            "fas icon-money-check-alt",
-            "fas icon-monument",
-            "fas icon-moon",
-            "fas icon-mortar-pestle",
-            "fas icon-mosque",
-            "fas icon-motorcycle",
-            "fas icon-mountain",
-            "fas icon-mouse",
-            "fas icon-mouse-pointer",
-            "fas icon-mug-hot",
-            "fas icon-music",
-            "fas icon-network-wired",
-            "fas icon-neuter",
-            "fas icon-newspaper",
-            "fas icon-not-equal",
-            "fas icon-notes-medical",
-            "fas icon-object-group",
-            "fas icon-object-ungroup",
-            "fas icon-oil-can",
-            "fas icon-om",
-            "fas icon-otter",
-            "fas icon-outdent",
-            "fas icon-pager",
-            "fas icon-paint-brush",
-            "fas icon-paint-roller",
-            "fas icon-palette",
-            "fas icon-pallet",
-            "fas icon-paper-plane",
-            "fas icon-paperclip",
-            "fas icon-parachute-box",
-            "fas icon-paragraph",
-            "fas icon-parking",
-            "fas icon-passport",
-            "fas icon-pastafarianism",
-            "fas icon-paste",
-            "fas icon-pause",
-            "fas icon-pause-circle",
-            "fas icon-paw",
-            "fas icon-peace",
-            "fas icon-pen",
-            "fas icon-pen-alt",
-            "fas icon-pen-fancy",
-            "fas icon-pen-nib",
-            "fas icon-pen-square",
-            "fas icon-pencil-alt",
-            "fas icon-pencil-ruler",
-            "fas icon-people-carry",
-            "fas icon-pepper-hot",
-            "fas icon-percent",
-            "fas icon-percentage",
-            "fas icon-person-booth",
-            "fas icon-phone",
-            "fas icon-phone-alt",
-            "fas icon-phone-slash",
-            "fas icon-phone-square",
-            "fas icon-phone-square-alt",
-            "fas icon-phone-volume",
-            "fas icon-photo-video",
-            "fas icon-piggy-bank",
-            "fas icon-pills",
-            "fas icon-pizza-slice",
-            "fas icon-place-of-worship",
-            "fas icon-plane",
-            "fas icon-plane-arrival",
-            "fas icon-plane-departure",
-            "fas icon-play",
-            "fas icon-play-circle",
-            "fas icon-plug",
-            "fas icon-plus",
-            "fas icon-plus-circle",
-            "fas icon-plus-square",
-            "fas icon-podcast",
-            "fas icon-poll",
-            "fas icon-poll-h",
-            "fas icon-poo",
-            "fas icon-poo-storm",
-            "fas icon-poop",
-            "fas icon-portrait",
-            "fas icon-pound-sign",
-            "fas icon-power-off",
-            "fas icon-pray",
-            "fas icon-praying-hands",
-            "fas icon-prescription",
-            "fas icon-prescription-bottle",
-            "fas icon-prescription-bottle-alt",
-            "fas icon-print",
-            "fas icon-procedures",
-            "fas icon-project-diagram",
-            "fas icon-puzzle-piece",
-            "fas icon-qrcode",
-            "fas icon-question",
-            "fas icon-question-circle",
-            "fas icon-quidditch",
-            "fas icon-quote-left",
-            "fas icon-quote-right",
-            "fas icon-quran",
-            "fas icon-radiation",
-            "fas icon-radiation-alt",
-            "fas icon-rainbow",
-            "fas icon-random",
-            "fas icon-receipt",
-            "fas icon-record-vinyl",
-            "fas icon-recycle",
-            "fas icon-redo",
-            "fas icon-redo-alt",
-            "fas icon-registered",
-            "fas icon-remove-format",
-            "fas icon-reply",
-            "fas icon-reply-all",
-            "fas icon-republican",
-            "fas icon-restroom",
-            "fas icon-retweet",
-            "fas icon-ribbon",
-            "fas icon-ring",
-            "fas icon-road",
-            "fas icon-robot",
-            "fas icon-rocket",
-            "fas icon-route",
-            "fas icon-rss",
-            "fas icon-rss-square",
-            "fas icon-ruble-sign",
-            "fas icon-ruler",
-            "fas icon-ruler-combined",
-            "fas icon-ruler-horizontal",
-            "fas icon-ruler-vertical",
-            "fas icon-running",
-            "fas icon-rupee-sign",
-            "fas icon-sad-cry",
-            "fas icon-sad-tear",
-            "fas icon-satellite",
-            "fas icon-satellite-dish",
-            "fas icon-save",
-            "fas icon-school",
-            "fas icon-screwdriver",
-            "fas icon-scroll",
-            "fas icon-sd-card",
-            "fas icon-search",
-            "fas icon-search-dollar",
-            "fas icon-search-location",
-            "fas icon-search-minus",
-            "fas icon-search-plus",
-            "fas icon-seedling",
-            "fas icon-server",
-            "fas icon-shapes",
-            "fas icon-share",
-            "fas icon-share-alt",
-            "fas icon-share-alt-square",
-            "fas icon-share-square",
-            "fas icon-shekel-sign",
-            "fas icon-shield-alt",
-            "fas icon-ship",
-            "fas icon-shipping-fast",
-            "fas icon-shoe-prints",
-            "fas icon-shopping-bag",
-            "fas icon-shopping-basket",
-            "fas icon-shopping-cart",
-            "fas icon-shower",
-            "fas icon-shuttle-van",
-            "fas icon-sign",
-            "fas icon-sign-in-alt",
-            "fas icon-sign-language",
-            "fas icon-sign-out-alt",
-            "fas icon-signal",
-            "fas icon-signature",
-            "fas icon-sim-card",
-            "fas icon-sitemap",
-            "fas icon-skating",
-            "fas icon-skiing",
-            "fas icon-skiing-nordic",
-            "fas icon-skull",
-            "fas icon-skull-crossbones",
-            "fas icon-slash",
-            "fas icon-sleigh",
-            "fas icon-sliders-h",
-            "fas icon-smile",
-            "fas icon-smile-beam",
-            "fas icon-smile-wink",
-            "fas icon-smog",
-            "fas icon-smoking",
-            "fas icon-smoking-ban",
-            "fas icon-sms",
-            "fas icon-snowboarding",
-            "fas icon-snowflake",
-            "fas icon-snowman",
-            "fas icon-snowplow",
-            "fas icon-socks",
-            "fas icon-solar-panel",
-            "fas icon-sort",
-            "fas icon-sort-alpha-down",
-            "fas icon-sort-alpha-down-alt",
-            "fas icon-sort-alpha-up",
-            "fas icon-sort-alpha-up-alt",
-            "fas icon-sort-amount-down",
-            "fas icon-sort-amount-down-alt",
-            "fas icon-sort-amount-up",
-            "fas icon-sort-amount-up-alt",
-            "fas icon-sort-down",
-            "fas icon-sort-numeric-down",
-            "fas icon-sort-numeric-down-alt",
-            "fas icon-sort-numeric-up",
-            "fas icon-sort-numeric-up-alt",
-            "fas icon-sort-up",
-            "fas icon-spa",
-            "fas icon-space-shuttle",
-            "fas icon-spell-check",
-            "fas icon-spider",
-            "fas icon-spinner",
-            "fas icon-splotch",
-            "fas icon-spray-can",
-            "fas icon-square",
-            "fas icon-square-full",
-            "fas icon-square-root-alt",
-            "fas icon-stamp",
-            "fas icon-star",
-            "fas icon-star-and-crescent",
-            "fas icon-star-half",
-            "fas icon-star-half-alt",
-            "fas icon-star-of-david",
-            "fas icon-star-of-life",
-            "fas icon-step-backward",
-            "fas icon-step-forward",
-            "fas icon-stethoscope",
-            "fas icon-sticky-note",
-            "fas icon-stop",
-            "fas icon-stop-circle",
-            "fas icon-stopwatch",
-            "fas icon-store",
-            "fas icon-store-alt",
-            "fas icon-stream",
-            "fas icon-street-view",
-            "fas icon-strikethrough",
-            "fas icon-stroopwafel",
-            "fas icon-subscript",
-            "fas icon-subway",
-            "fas icon-suitcase",
-            "fas icon-suitcase-rolling",
-            "fas icon-sun",
-            "fas icon-superscript",
-            "fas icon-surprise",
-            "fas icon-swatchbook",
-            "fas icon-swimmer",
-            "fas icon-swimming-pool",
-            "fas icon-synagogue",
-            "fas icon-sync",
-            "fas icon-sync-alt",
-            "fas icon-syringe",
-            "fas icon-table",
-            "fas icon-table-tennis",
-            "fas icon-tablet",
-            "fas icon-tablet-alt",
-            "fas icon-tablets",
-            "fas icon-tachometer-alt",
-            "fas icon-tag",
-            "fas icon-tags",
-            "fas icon-tape",
-            "fas icon-tasks",
-            "fas icon-taxi",
-            "fas icon-teeth",
-            "fas icon-teeth-open",
-            "fas icon-temperature-high",
-            "fas icon-temperature-low",
-            "fas icon-tenge",
-            "fas icon-terminal",
-            "fas icon-text-height",
-            "fas icon-text-width",
-            "fas icon-th",
-            "fas icon-th-large",
-            "fas icon-th-list",
-            "fas icon-theater-masks",
-            "fas icon-thermometer",
-            "fas icon-thermometer-empty",
-            "fas icon-thermometer-full",
-            "fas icon-thermometer-half",
-            "fas icon-thermometer-quarter",
-            "fas icon-thermometer-three-quarters",
-            "fas icon-thumbs-down",
-            "fas icon-thumbs-up",
-            "fas icon-thumbtack",
-            "fas icon-ticket-alt",
-            "fas icon-times",
-            "fas icon-times-circle",
-            "fas icon-tint",
-            "fas icon-tint-slash",
-            "fas icon-tired",
-            "fas icon-toggle-off",
-            "fas icon-toggle-on",
-            "fas icon-toilet",
-            "fas icon-toilet-paper",
-            "fas icon-toolbox",
-            "fas icon-tools",
-            "fas icon-tooth",
-            "fas icon-torah",
-            "fas icon-torii-gate",
-            "fas icon-tractor",
-            "fas icon-trademark",
-            "fas icon-traffic-light",
-            "fas icon-trailer",
-            "fas icon-train",
-            "fas icon-tram",
-            "fas icon-transgender",
-            "fas icon-transgender-alt",
-            "fas icon-trash",
-            "fas icon-trash-alt",
-            "fas icon-trash-restore",
-            "fas icon-trash-restore-alt",
-            "fas icon-tree",
-            "fas icon-trophy",
-            "fas icon-truck",
-            "fas icon-truck-loading",
-            "fas icon-truck-monster",
-            "fas icon-truck-moving",
-            "fas icon-truck-pickup",
-            "fas icon-tshirt",
-            "fas icon-tty",
-            "fas icon-tv",
-            "fas icon-umbrella",
-            "fas icon-umbrella-beach",
-            "fas icon-underline",
-            "fas icon-undo",
-            "fas icon-undo-alt",
-            "fas icon-universal-access",
-            "fas icon-university",
-            "fas icon-unlink",
-            "fas icon-unlock",
-            "fas icon-unlock-alt",
-            "fas icon-upload",
-            "fas icon-user",
-            "fas icon-user-alt",
-            "fas icon-user-alt-slash",
-            "fas icon-user-astronaut",
-            "fas icon-user-check",
-            "fas icon-user-circle",
-            "fas icon-user-clock",
-            "fas icon-user-cog",
-            "fas icon-user-edit",
-            "fas icon-user-friends",
-            "fas icon-user-graduate",
-            "fas icon-user-injured",
-            "fas icon-user-lock",
-            "fas icon-user-md",
-            "fas icon-user-minus",
-            "fas icon-user-ninja",
-            "fas icon-user-nurse",
-            "fas icon-user-plus",
-            "fas icon-user-secret",
-            "fas icon-user-shield",
-            "fas icon-user-slash",
-            "fas icon-user-tag",
-            "fas icon-user-tie",
-            "fas icon-user-times",
-            "fas icon-users",
-            "fas icon-users-cog",
-            "fas icon-utensil-spoon",
-            "fas icon-utensils",
-            "fas icon-vector-square",
-            "fas icon-venus",
-            "fas icon-venus-double",
-            "fas icon-venus-mars",
-            "fas icon-vial",
-            "fas icon-vials",
-            "fas icon-video",
-            "fas icon-video-slash",
-            "fas icon-vihara",
-            "fas icon-voicemail",
-            "fas icon-volleyball-ball",
-            "fas icon-volume-down",
-            "fas icon-volume-mute",
-            "fas icon-volume-off",
-            "fas icon-volume-up",
-            "fas icon-vote-yea",
-            "fas icon-vr-cardboard",
-            "fas icon-walking",
-            "fas icon-wallet",
-            "fas icon-warehouse",
-            "fas icon-water",
-            "fas icon-wave-square",
-            "fas icon-weight",
-            "fas icon-weight-hanging",
-            "fas icon-wheelchair",
-            "fas icon-wifi",
-            "fas icon-wind",
-            "fas icon-window-close",
-            "fas icon-window-maximize",
-            "fas icon-window-minimize",
-            "fas icon-window-restore",
-            "fas icon-wine-bottle",
-            "fas icon-wine-glass",
-            "fas icon-wine-glass-alt",
-            "fas icon-won-sign",
-            "fas icon-wrench",
-            "fas icon-x-ray",
-            "fas icon-yen-sign",
-            "fas icon-yin-yang"
-        ]
-    ],
-    [
-        "id" => "fab",
-        "title" => "Font Awesome Brands",
-        "prefix" => "fab icon-",
-        "listicon" => "fab icon-font-awesome-flag",
-        "icons" => [
-            "fab icon-500px",
-            "fab icon-accessible-icon",
-            "fab icon-accusoft",
-            "fab icon-adn",
-            "fab icon-adversal",
-            "fab icon-affiliatetheme",
-            "fab icon-airbnb",
-            "fab icon-algolia",
-            "fab icon-alipay",
-            "fab icon-amazon",
-            "fab icon-amazon-pay",
-            "fab icon-amilia",
-            "fab icon-android",
-            "fab icon-angellist",
-            "fab icon-angrycreative",
-            "fab icon-angular",
-            "fab icon-app-store",
-            "fab icon-app-store-ios",
-            "fab icon-apper",
-            "fab icon-apple",
-            "fab icon-apple-pay",
-            "fab icon-artstation",
-            "fab icon-asymmetrik",
-            "fab icon-atlassian",
-            "fab icon-audible",
-            "fab icon-autoprefixer",
-            "fab icon-avianex",
-            "fab icon-aviato",
-            "fab icon-aws",
-            "fab icon-bandcamp",
-            "fab icon-battle-net",
-            "fab icon-behance",
-            "fab icon-behance-square",
-            "fab icon-bimobject",
-            "fab icon-bitbucket",
-            "fab icon-bitcoin",
-            "fab icon-bity",
-            "fab icon-black-tie",
-            "fab icon-blackberry",
-            "fab icon-blogger",
-            "fab icon-blogger-b",
-            "fab icon-bluetooth",
-            "fab icon-bluetooth-b",
-            "fab icon-bootstrap",
-            "fab icon-btc",
-            "fab icon-buffer",
-            "fab icon-buromobelexperte",
-            "fab icon-buy-n-large",
-            "fab icon-buysellads",
-            "fab icon-canadian-maple-leaf",
-            "fab icon-cc-amazon-pay",
-            "fab icon-cc-amex",
-            "fab icon-cc-apple-pay",
-            "fab icon-cc-diners-club",
-            "fab icon-cc-discover",
-            "fab icon-cc-jcb",
-            "fab icon-cc-mastercard",
-            "fab icon-cc-paypal",
-            "fab icon-cc-stripe",
-            "fab icon-cc-visa",
-            "fab icon-centercode",
-            "fab icon-centos",
-            "fab icon-chrome",
-            "fab icon-chromecast",
-            "fab icon-cloudscale",
-            "fab icon-cloudsmith",
-            "fab icon-cloudversify",
-            "fab icon-codepen",
-            "fab icon-codiepie",
-            "fab icon-confluence",
-            "fab icon-connectdevelop",
-            "fab icon-contao",
-            "fab icon-cotton-bureau",
-            "fab icon-cpanel",
-            "fab icon-creative-commons",
-            "fab icon-creative-commons-by",
-            "fab icon-creative-commons-nc",
-            "fab icon-creative-commons-nc-eu",
-            "fab icon-creative-commons-nc-jp",
-            "fab icon-creative-commons-nd",
-            "fab icon-creative-commons-pd",
-            "fab icon-creative-commons-pd-alt",
-            "fab icon-creative-commons-remix",
-            "fab icon-creative-commons-sa",
-            "fab icon-creative-commons-sampling",
-            "fab icon-creative-commons-sampling-plus",
-            "fab icon-creative-commons-share",
-            "fab icon-creative-commons-zero",
-            "fab icon-critical-role",
-            "fab icon-css3",
-            "fab icon-css3-alt",
-            "fab icon-cuttlefish",
-            "fab icon-d-and-d",
-            "fab icon-d-and-d-beyond",
-            "fab icon-dashcube",
-            "fab icon-delicious",
-            "fab icon-deploydog",
-            "fab icon-deskpro",
-            "fab icon-dev",
-            "fab icon-deviantart",
-            "fab icon-dhl",
-            "fab icon-diaspora",
-            "fab icon-digg",
-            "fab icon-digital-ocean",
-            "fab icon-discord",
-            "fab icon-discourse",
-            "fab icon-dochub",
-            "fab icon-docker",
-            "fab icon-draft2digital",
-            "fab icon-dribbble",
-            "fab icon-dribbble-square",
-            "fab icon-dropbox",
-            "fab icon-drupal",
-            "fab icon-dyalog",
-            "fab icon-earlybirds",
-            "fab icon-ebay",
-            "fab icon-edge",
-            "fab icon-elementor",
-            "fab icon-ello",
-            "fab icon-ember",
-            "fab icon-empire",
-            "fab icon-envira",
-            "fab icon-erlang",
-            "fab icon-ethereum",
-            "fab icon-etsy",
-            "fab icon-evernote",
-            "fab icon-expeditedssl",
-            "fab icon-facebook",
-            "fab icon-facebook-f",
-            "fab icon-facebook-messenger",
-            "fab icon-facebook-square",
-            "fab icon-fantasy-flight-games",
-            "fab icon-fedex",
-            "fab icon-fedora",
-            "fab icon-figma",
-            "fab icon-firefox",
-            "fab icon-firefox-browser",
-            "fab icon-first-order",
-            "fab icon-first-order-alt",
-            "fab icon-firstdraft",
-            "fab icon-flickr",
-            "fab icon-flipboard",
-            "fab icon-fly",
-            "fab icon-font-awesome",
-            "fab icon-font-awesome-alt",
-            "fab icon-font-awesome-flag",
-            "fab icon-fonticons",
-            "fab icon-fonticons-fi",
-            "fab icon-fort-awesome",
-            "fab icon-fort-awesome-alt",
-            "fab icon-forumbee",
-            "fab icon-foursquare",
-            "fab icon-free-code-camp",
-            "fab icon-freebsd",
-            "fab icon-fulcrum",
-            "fab icon-galactic-republic",
-            "fab icon-galactic-senate",
-            "fab icon-get-pocket",
-            "fab icon-gg",
-            "fab icon-gg-circle",
-            "fab icon-git",
-            "fab icon-git-alt",
-            "fab icon-git-square",
-            "fab icon-github",
-            "fab icon-github-alt",
-            "fab icon-github-square",
-            "fab icon-gitkraken",
-            "fab icon-gitlab",
-            "fab icon-gitter",
-            "fab icon-glide",
-            "fab icon-glide-g",
-            "fab icon-gofore",
-            "fab icon-goodreads",
-            "fab icon-goodreads-g",
-            "fab icon-google",
-            "fab icon-google-drive",
-            "fab icon-google-play",
-            "fab icon-google-plus",
-            "fab icon-google-plus-g",
-            "fab icon-google-plus-square",
-            "fab icon-google-wallet",
-            "fab icon-gratipay",
-            "fab icon-grav",
-            "fab icon-gripfire",
-            "fab icon-grunt",
-            "fab icon-gulp",
-            "fab icon-hacker-news",
-            "fab icon-hacker-news-square",
-            "fab icon-hackerrank",
-            "fab icon-hips",
-            "fab icon-hire-a-helper",
-            "fab icon-hooli",
-            "fab icon-hornbill",
-            "fab icon-hotjar",
-            "fab icon-houzz",
-            "fab icon-html5",
-            "fab icon-hubspot",
-            "fab icon-ideal",
-            "fab icon-imdb",
-            "fab icon-instagram",
-            "fab icon-intercom",
-            "fab icon-internet-explorer",
-            "fab icon-invision",
-            "fab icon-ioxhost",
-            "fab icon-itch-io",
-            "fab icon-itunes",
-            "fab icon-itunes-note",
-            "fab icon-java",
-            "fab icon-jedi-order",
-            "fab icon-jenkins",
-            "fab icon-jira",
-            "fab icon-joget",
-            "fab icon-joomla",
-            "fab icon-js",
-            "fab icon-js-square",
-            "fab icon-jsfiddle",
-            "fab icon-kaggle",
-            "fab icon-keybase",
-            "fab icon-keycdn",
-            "fab icon-kickstarter",
-            "fab icon-kickstarter-k",
-            "fab icon-korvue",
-            "fab icon-laravel",
-            "fab icon-lastfm",
-            "fab icon-lastfm-square",
-            "fab icon-leanpub",
-            "fab icon-less",
-            "fab icon-line",
-            "fab icon-linkedin",
-            "fab icon-linkedin-in",
-            "fab icon-linode",
-            "fab icon-linux",
-            "fab icon-lyft",
-            "fab icon-magento",
-            "fab icon-mailchimp",
-            "fab icon-mandalorian",
-            "fab icon-markdown",
-            "fab icon-mastodon",
-            "fab icon-maxcdn",
-            "fab icon-mdb",
-            "fab icon-medapps",
-            "fab icon-medium",
-            "fab icon-medium-m",
-            "fab icon-medrt",
-            "fab icon-meetup",
-            "fab icon-megaport",
-            "fab icon-mendeley",
-            "fab icon-microblog",
-            "fab icon-microsoft",
-            "fab icon-mix",
-            "fab icon-mixcloud",
-            "fab icon-mizuni",
-            "fab icon-modx",
-            "fab icon-monero",
-            "fab icon-napster",
-            "fab icon-neos",
-            "fab icon-nimblr",
-            "fab icon-node",
-            "fab icon-node-js",
-            "fab icon-npm",
-            "fab icon-ns8",
-            "fab icon-nutritionix",
-            "fab icon-odnoklassniki",
-            "fab icon-odnoklassniki-square",
-            "fab icon-old-republic",
-            "fab icon-opencart",
-            "fab icon-openid",
-            "fab icon-opera",
-            "fab icon-optin-monster",
-            "fab icon-orcid",
-            "fab icon-osi",
-            "fab icon-page4",
-            "fab icon-pagelines",
-            "fab icon-palfed",
-            "fab icon-patreon",
-            "fab icon-paypal",
-            "fab icon-periscope",
-            "fab icon-phabricator",
-            "fab icon-phoenix-framework",
-            "fab icon-phoenix-squadron",
-            "fab icon-php",
-            "fab icon-pied-piper",
-            "fab icon-pied-piper-alt",
-            "fab icon-pied-piper-hat",
-            "fab icon-pied-piper-pp",
-            "fab icon-pied-piper-square",
-            "fab icon-pinterest",
-            "fab icon-pinterest-p",
-            "fab icon-pinterest-square",
-            "fab icon-playstation",
-            "fab icon-product-hunt",
-            "fab icon-pushed",
-            "fab icon-python",
-            "fab icon-qq",
-            "fab icon-quinscape",
-            "fab icon-quora",
-            "fab icon-r-project",
-            "fab icon-raspberry-pi",
-            "fab icon-ravelry",
-            "fab icon-react",
-            "fab icon-reacteurope",
-            "fab icon-readme",
-            "fab icon-rebel",
-            "fab icon-red-river",
-            "fab icon-reddit",
-            "fab icon-reddit-alien",
-            "fab icon-reddit-square",
-            "fab icon-redhat",
-            "fab icon-renren",
-            "fab icon-replyd",
-            "fab icon-researchgate",
-            "fab icon-resolving",
-            "fab icon-rev",
-            "fab icon-rocketchat",
-            "fab icon-rockrms",
-            "fab icon-safari",
-            "fab icon-salesforce",
-            "fab icon-sass",
-            "fab icon-schlix",
-            "fab icon-scribd",
-            "fab icon-searchengin",
-            "fab icon-sellcast",
-            "fab icon-sellsy",
-            "fab icon-servicestack",
-            "fab icon-shirtsinbulk",
-            "fab icon-shopware",
-            "fab icon-simplybuilt",
-            "fab icon-sistrix",
-            "fab icon-sith",
-            "fab icon-sketch",
-            "fab icon-skyatlas",
-            "fab icon-skype",
-            "fab icon-slack",
-            "fab icon-slack-hash",
-            "fab icon-slideshare",
-            "fab icon-snapchat",
-            "fab icon-snapchat-ghost",
-            "fab icon-snapchat-square",
-            "fab icon-soundcloud",
-            "fab icon-sourcetree",
-            "fab icon-speakap",
-            "fab icon-speaker-deck",
-            "fab icon-spotify",
-            "fab icon-squarespace",
-            "fab icon-stack-exchange",
-            "fab icon-stack-overflow",
-            "fab icon-stackpath",
-            "fab icon-staylinked",
-            "fab icon-steam",
-            "fab icon-steam-square",
-            "fab icon-steam-symbol",
-            "fab icon-sticker-mule",
-            "fab icon-strava",
-            "fab icon-stripe",
-            "fab icon-stripe-s",
-            "fab icon-studiovinari",
-            "fab icon-stumbleupon",
-            "fab icon-stumbleupon-circle",
-            "fab icon-superpowers",
-            "fab icon-supple",
-            "fab icon-suse",
-            "fab icon-swift",
-            "fab icon-symfony",
-            "fab icon-teamspeak",
-            "fab icon-telegram",
-            "fab icon-telegram-plane",
-            "fab icon-tencent-weibo",
-            "fab icon-the-red-yeti",
-            "fab icon-themeco",
-            "fab icon-themeisle",
-            "fab icon-think-peaks",
-            "fab icon-trade-federation",
-            "fab icon-trello",
-            "fab icon-tumblr",
-            "fab icon-tumblr-square",
-            "fab icon-twitch",
-            "fab icon-twitter",
-            "fab icon-twitter-square",
-            "fab icon-typo3",
-            "fab icon-uber",
-            "fab icon-ubuntu",
-            "fab icon-uikit",
-            "fab icon-umbraco",
-            "fab icon-uniregistry",
-            "fab icon-unity",
-            "fab icon-untappd",
-            "fab icon-ups",
-            "fab icon-usb",
-            "fab icon-usps",
-            "fab icon-ussunnah",
-            "fab icon-vaadin",
-            "fab icon-viacoin",
-            "fab icon-viadeo",
-            "fab icon-viadeo-square",
-            "fab icon-viber",
-            "fab icon-vimeo",
-            "fab icon-vimeo-square",
-            "fab icon-vimeo-v",
-            "fab icon-vine",
-            "fab icon-vk",
-            "fab icon-vnv",
-            "fab icon-vuejs",
-            "fab icon-waze",
-            "fab icon-weebly",
-            "fab icon-weibo",
-            "fab icon-weixin",
-            "fab icon-whatsapp",
-            "fab icon-whatsapp-square",
-            "fab icon-whmcs",
-            "fab icon-wikipedia-w",
-            "fab icon-windows",
-            "fab icon-wix",
-            "fab icon-wizards-of-the-coast",
-            "fab icon-wolf-pack-battalion",
-            "fab icon-wordpress",
-            "fab icon-wordpress-simple",
-            "fab icon-wpbeginner",
-            "fab icon-wpexplorer",
-            "fab icon-wpforms",
-            "fab icon-wpressr",
-            "fab icon-xbox",
-            "fab icon-xing",
-            "fab icon-xing-square",
-            "fab icon-y-combinator",
-            "fab icon-yahoo",
-            "fab icon-yammer",
-            "fab icon-yandex",
-            "fab icon-yandex-international",
-            "fab icon-yarn",
-            "fab icon-yelp",
-            "fab icon-yoast",
-            "fab icon-youtube",
-            "fab icon-youtube-square",
-            "fab icon-zhihu"
-        ]
-    ]
-];
