+++ modules/backend/formwidgets/iconpicker/assets/src/components/IconPicker.vue
