+++ modules/backend/formwidgets/iconpicker/assets/src/iconpicker.js
@@ -1,18 +1,18 @@
 import { createApp } from 'vue'
-import App from './components/App.vue'
+import IconPicker from './components/IconPicker.vue'
 
 window.addEventListener('load', () => {
     const elements = document.querySelectorAll("[data-control='iconpicker']:not([data-v-app])");
     elements.forEach(element => {
         element.querySelector(".input-group").addEventListener("click", () => {
-            snowboard.request(null, "onLoadIconLibrary", {
+            snowboard.request(null, element.dataset.alias + "::onLoadIconLibrary", {
                 success: (data) => {
-                    let app = createApp(App, {
+                    let iconPicker = createApp(IconPicker, {
                         ...element.dataset,
                         fontLibraries: JSON.parse(data.result)
                     }).mount(element);
 
-                    app.togglePicker();
+                    iconPicker.togglePicker();
                 },
             })
         });
