+++ modules/backend/formwidgets/iconpicker/partials/_iconpicker.php
@@ -9,12 +9,17 @@
          data-label="Icon Picker"
          data-prop-value="<?= e($field->getLoadValue()) ?>"
          data-name="<?= $field->getFieldName() ?>"
+         data-alias="<?= $field->alias ?>"
     >
         <div class="input-group">
             <span class="input-group-addon" style="cursor: pointer">
-                <i class="<?= $field->getLoadValue() ? e($field->getLoadValue()) : 'far icon-address-book' ?>"></i>
+                <i class="<?= $field->getLoadValue() ? e($field->getLoadValue()) : ($field->config->default ?? 'far icon-address-book') ?>"></i>
             </span>
-            <input type="text" class="form-control" value="<?= $field->getLoadValue() ? e($field->getLoadValue()) : 'far icon-address-book' ?>">
+            <input type="text"
+                   class="form-control"
+                   name="<?= $field->getFieldName() ?>"
+                   value="<?= $field->getLoadValue() ? e($field->getLoadValue()) : ($field->config->default ?? 'far icon-address-book') ?>"
+            >
         </div>
     </div>
 <?php endif ?>
